+++ src/cli/runtime/index.ts
@@ -1,0 +1,9 @@
+export { RuntimeClient } from './client'
+export { getDefaultUserDataPath } from './metadata'
+export {
+  RuntimeClientError,
+  RuntimeRpcFailureError,
+  type RuntimeRpcFailure,
+  type RuntimeRpcResponse,
+  type RuntimeRpcSuccess
+} from './types'
